+++ lib/src/world/world.dart
@@ -17,6 +17,7 @@ class World {
   }
 
   void process({Duration delta = const Duration(milliseconds: 16)}) {
+    _entityManager.processCreationQueue();
     for (var system in _systems) {
       system.process(delta);
       _entityManager.processDeletionQueue();